CO: First pass at 2017 bill scraper
